book-sync/requirements.txt
mysql-connector-python>=9.0
pymongo>=4.6
